requirements.txt
# Core
python-dotenv

# Hugging Face datasets
datasets

# Progress bars
tqdm

# LangChain
langchain-core
langchain-community
langchain-openai
langchain-google-genai

# Isaacus integration
langchain-isaacus

# FAISS (vector index backend)
faiss-cpu